+++ src/sentry/ingest/consumer/processors.py
@@ -14,7 +14,7 @@ from sentry import eventstore, features
 from sentry.attachments import CachedAttachment, attachment_cache
 from sentry.event_manager import EventManager, save_attachment
 from sentry.eventstore.processing import event_processing_store, transaction_processing_store
-from sentry.feedback.usecases.create_feedback import FeedbackCreationSource, is_in_feedback_denylist
+from sentry.feedback.lib.utils import FeedbackCreationSource, is_in_feedback_denylist
 from sentry.ingest.types import ConsumerType
 from sentry.ingest.userreport import Conflict, save_userreport
 from sentry.killswitches import killswitch_matches_context
