+++ src/sentry/web/frontend/error_page_embed.py
@@ -12,7 +12,8 @@ from django.views.decorators.csrf import csrf_exempt
 from django.views.generic import View
 
 from sentry import eventstore
-from sentry.feedback.usecases.create_feedback import FeedbackCreationSource, shim_to_feedback
+from sentry.feedback.lib.utils import FeedbackCreationSource
+from sentry.feedback.usecases.create_feedback import shim_to_feedback
 from sentry.models.options.project_option import ProjectOption
 from sentry.models.project import Project
 from sentry.models.projectkey import ProjectKey
