+++ tests/sentry/api/endpoints/test_organization_feedback_summary.py
@@ -3,7 +3,8 @@ from unittest.mock import patch
 
 from django.urls import reverse
 
-from sentry.feedback.usecases.create_feedback import FeedbackCreationSource, create_feedback_issue
+from sentry.feedback.lib.utils import FeedbackCreationSource
+from sentry.feedback.usecases.create_feedback import create_feedback_issue
 from sentry.testutils.cases import APITestCase
 from sentry.testutils.pytest.fixtures import django_db_all
 from sentry.testutils.silo import region_silo_test
