+++ tests/sentry/issues/endpoints/test_organization_group_index.py
@@ -12,7 +12,8 @@ from django.urls import reverse
 from django.utils import timezone
 
 from sentry import options
-from sentry.feedback.usecases.create_feedback import FeedbackCreationSource, create_feedback_issue
+from sentry.feedback.lib.utils import FeedbackCreationSource
+from sentry.feedback.usecases.create_feedback import create_feedback_issue
 from sentry.integrations.models.external_issue import ExternalIssue
 from sentry.integrations.models.organization_integration import OrganizationIntegration
 from sentry.issues.grouptype import (
